Clean up of tritext.
